Add docu about texture format, test and fix it

+++ src/ArchivItem_Sound_XMidi.cpp
@@ -186,9 +186,9 @@ int baseArchivItem_Sound_XMidi::write(std::ostream& file) const
     {
         uint32_t trackLength = 4 + sizeof(uint32_t) + 4; // XMID EVNT len
         const XMIDI_Track& track = tracklist[i];
-        trackLength += track.getData().size();
+        trackLength += static_cast<uint32_t>(track.getData().size());
         if(!track.getTimbres().empty())
-            trackLength += 4 + sizeof(uint32_t) + sizeof(uint16_t) + track.getTimbres().size() * 2;
+            trackLength += 4 + sizeof(uint32_t) + sizeof(uint16_t) + static_cast<uint32_t>(track.getTimbres().size()) * 2;
         trackLengths[i] = trackLength;
         len += 4 + sizeof(uint32_t) + trackLength; // FORM len
     }
