+++ src/PixelBufferARGB.h
@@ -30,7 +30,7 @@ namespace libsiedler2
     {
     public:
         PixelBufferARGB(){}
-        PixelBufferARGB(uint16_t width, uint16_t height): PixelBuffer<uint8_t>(width, height, 0){}
+        PixelBufferARGB(uint16_t width, uint16_t height): PixelBuffer<uint32_t>(width, height, 0){}
         ColorARGB get(uint16_t x, uint16_t y) const;
         ColorARGB get(uint32_t idx) const;
         void set(uint16_t x, uint16_t y, ColorARGB clr);
